added locking for concurrent access of map

+++ engine/storage.go
@@ -1,6 +1,7 @@
 package engine
 
 import (
+	"sync"
 	"universum/config"
 	"universum/consts"
 	"universum/engine/entity"
@@ -9,8 +10,13 @@ import (
 )
 
 var memoryStore map[unsafe.Pointer]*entity.Record
+var memStoreMutex sync.RWMutex
+
 var expirationMap map[*entity.Record]uint32
+var expirationMutex sync.RWMutex
+
 var keypool map[string]unsafe.Pointer
+var keypoolMutex sync.RWMutex
 
 type Storage struct {
 }
@@ -22,12 +28,18 @@ func (s *Storage) Initialize() {
 }
 
 func (s *Storage) Exists(key string) (bool, uint32) {
+	keypoolMutex.RLock()
 	pointer, ok := keypool[key]
+	keypoolMutex.RUnlock()
+
 	if !ok {
 		return false, consts.CRC_RECORD_NOT_FOUND
 	}
 
+	memStoreMutex.RLock()
 	record, ok := memoryStore[pointer]
+	memStoreMutex.RUnlock()
+
 	if ok {
 		expired := isExpiredRecord(record)
 
@@ -43,12 +55,18 @@ func (s *Storage) Exists(key string) (bool, uint32) {
 }
 
 func (s *Storage) Get(key string) (*entity.Record, uint32) {
+	keypoolMutex.RLock()
 	pointer, ok := keypool[key]
+	keypoolMutex.RUnlock()
+
 	if !ok {
 		return nil, consts.CRC_RECORD_NOT_FOUND
 	}
 
+	memStoreMutex.RLock()
 	record := memoryStore[pointer]
+	memStoreMutex.RUnlock()
+
 	if record == nil {
 		return nil, consts.CRC_RECORD_NOT_FOUND
 	}
@@ -63,20 +81,29 @@ func (s *Storage) Get(key string) (*entity.Record, uint32) {
 }
 
 func (s *Storage) Set(key string, value interface{}, ttl uint32) (bool, uint32) {
-	// @TODO put some eviction logic here
-
+	//TriggerPeriodicExpiredRecordCleaup()
 	record := entity.NewRecord(value, GetTypeEncoding(value), utils.GetCurrentEPochTime())
 
+	keypoolMutex.RLock()
 	pointer, ok := keypool[key]
+	keypoolMutex.RUnlock()
+
 	if !ok {
-		keypool[key] = unsafe.Pointer(&key)
 		pointer = unsafe.Pointer(&key)
+
+		keypoolMutex.Lock()
+		keypool[key] = pointer
+		keypoolMutex.Unlock()
 	}
 
+	memStoreMutex.Lock()
 	memoryStore[pointer] = record
+	memStoreMutex.Unlock()
 
 	if ttl > 0 {
+		expirationMutex.Lock()
 		expirationMap[record] = utils.GetCurrentEPochTime() + ttl
+		expirationMutex.Unlock()
 	}
 
 	return true, consts.CRC_RECORD_UPDATED
@@ -87,13 +114,20 @@ func (s *Storage) Delete(key string) (bool, uint32) {
 		return true, consts.CRC_RECORD_DELETED
 	}
 
-	pointer := keypool[key]
-	record := memoryStore[pointer]
+	keypoolMutex.RLock()
+	pointer, ok := keypool[key]
+	keypoolMutex.RUnlock()
 
-	deleted := deleteByPointer(record, pointer)
+	if ok {
+		memStoreMutex.RLock()
+		record := memoryStore[pointer]
+		memStoreMutex.RUnlock()
 
-	if deleted {
-		return true, consts.CRC_RECORD_DELETED
+		deleted := deleteByPointer(record, pointer)
+
+		if deleted {
+			return true, consts.CRC_RECORD_DELETED
+		}
 	}
 
 	return false, consts.CRC_RECORD_NOT_DELETED
@@ -114,7 +148,11 @@ func (s *Storage) IncrDecrInteger(key string, offset int64, isIncr bool) (int64,
 			}
 
 			var ttl uint32 = 0
+
+			expirationMutex.RLock()
 			expiry, ok := expirationMap[record]
+			expirationMutex.RUnlock()
+
 			if ok {
 				ttl = expiry - utils.GetCurrentEPochTime()
 			}
@@ -140,9 +178,17 @@ func (s *Storage) Append(key string, offset int64) (*entity.Record, uint32) {
 // ------------------- Internal Functions -------------------
 
 func deleteByPointer(record *entity.Record, pointer unsafe.Pointer) bool {
+	memStoreMutex.Lock()
 	delete(memoryStore, pointer)
+	memStoreMutex.Unlock()
+
+	keypoolMutex.Lock()
 	delete(keypool, *((*string)(pointer)))
+	keypoolMutex.Unlock()
+
+	expirationMutex.Lock()
 	delete(expirationMap, record)
+	expirationMutex.Unlock()
 
 	return true
 }
